+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - [#136](https://github.com/jaunt-lang/jaunt/pull/136) Support `^:deprecated` annotations on single fn arities (@arrdem).
 - [#138](https://github.com/jaunt-lang/jaunt/pull/138) Roll back deprecation of `clojure-version`, `*clojure-version*` (@arrdem).
 - [#135](https://github.com/jaunt-lang/jaunt/pull/134) Make `get` throw on non-associative arguments (@arrdem).
-- [#164](https://github.com/jaunt-lang/jaunt/pull/164) Add `defn` style docstring and attr-map support to `defonce`
+- [#164](https://github.com/jaunt-lang/jaunt/pull/164) Add `defn` style docstring and attr-map support to `defonce` (@arrdem).
+- [#155](https://github.com/jaunt-lang/jaunt/pull/155) Add `defn` style docstring and attr-map support to `defprotocol` (@arrdem).
 
 ## Jaunt 0.2.1